requirements.txt
Flask==2.3.2
Flask-Cors==3.1.1
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python==9.1.0
